CreateArrayOfDistinctIntegerInRandomOrder added  from source comment

+++ Code4Fun/Code4Fun.cs
@@ -30,40 +30,8 @@ namespace Code4Fun
             return ringBuffer[counter % index];
         }
 
-        //public static Int32[] CreateArrayOfDistinctIntegerInRandomOrder(Int32 size)
-        //{
-        //    if (size < 2)
-        //    {
-        //        throw new ArgumentOutOfRangeException("Size must be greater than one");
-        //    }
-
-        //    var sortedListOfInteger = new List<int>(size);
-        //    for (var n = 1; n <= size; n++ )
-        //    {
-        //        sortedListOfInteger.Add(n);
-        //    }
-
-        //    var arrayOfDistinctIntegerInRandomOrder = new Int32[size];
-        //    var random = new Random();
-
-        //    for (var n = 0; n < size; n++)
-        //    {
-        //        var sortedListOfIntegerRandomIndex = 
-        //            random.Next(sortedListOfInteger.Count);
-
-        //        arrayOfDistinctIntegerInRandomOrder[n] = 
-        //            sortedListOfInteger[sortedListOfIntegerRandomIndex];
-
-        //        sortedListOfInteger[sortedListOfIntegerRandomIndex] =
-        //           sortedListOfInteger[sortedListOfInteger.Count - 1];
-
-        //        sortedListOfInteger.RemoveAt(sortedListOfInteger.Count - 1);
-        //    }
-
-        //    return arrayOfDistinctIntegerInRandomOrder;
-        //}
-
-        public static IEnumerable<int> CreateArrayOfDistinctIntegerInRandomOrder(Int32 size)
+        //from http://stackoverflow.com/questions/464476/generating-shuffled-range-using-a-prng-rather-than-shuffling#465473
+        public static IEnumerable<Int32> CreateArrayOfDistinctIntegerInRandomOrder(Int32 size)
         {
             if (size < 2)
             {
@@ -72,16 +40,16 @@ namespace Code4Fun
 
             Random r = new Random();
             int M = Code4Fun.NextLargestPowerOfTwo(size);
-            int c = r.Next(M / 2) * 2 + 1; // make c any odd number between 0 and M
-            int a = r.Next(M / 4) * 4 + 1; // M = 2^m, so make (a-1) divisible by all prime factors, and 4
+            int c = r.Next(M / 2) * 2 + 1; 
+            int a = r.Next(M / 4) * 4 + 1; 
 
             int start = r.Next(M);
-            int x = start;
+            Int64 x = start;
             do
             {
-                x = (a * x + c) % M;
-                if (x < size )
-                    yield return x + 1;
+                x = ((a * x + c) % M);
+                if (x < size)
+                    yield return (Int32)x + 1;
             } while (x != start);
         }
 
@@ -94,6 +62,5 @@ namespace Code4Fun
             n |= (n >> 16);
             return (n + 1);
         }
-
     }
 }
